+++ CAP25/backend1/solver.py
